// lib/api/generated/schemas/adminSortDtoOrder.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * Textbooked Backend Admin API
 * Backend MVP API for textbook study planning and grading.
 * OpenAPI spec version: 1.0.0
 */

export type AdminSortDtoOrder = typeof AdminSortDtoOrder[keyof typeof AdminSortDtoOrder];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AdminSortDtoOrder = {
  ASC: 'ASC',
  DESC: 'DESC',
} as const;